refactor: Move files related to ideascale importer to a new package
